+++ client/src/components/svgs/index.ts
@@ -4,5 +4,5 @@ export {default as EditIcon} from './editIcon'
 export {default as ErrorIcon} from './errorIcon'
 export {default as KebabMenuIcon} from './kebabMenuIcon'
 export {default as ExportIcon} from './exportIcon'
-export {default as SpeakerIcon} from './speakerIcon'
-export {default as questionMarkIcon} from './questionMarkIcon'
+export {default as SpeakerIcon} from './sspeakerIcon'
+export {default as QuestionMarkIcon} from './qquestionMarkIcon'
