+++ client/src/components/svgs/sspeakerIcon.tsx
